カウンター画面 カウンター処理をコンポーザブルに抽出
